requirements.txt
flask>=3.0
flask-cors>=4.0
gunicorn>=22.0
requests
python-dotenv
cryptography
kalshi_python
